requirements.txt
coppeliasim_zmqremoteapi_client~=2.0.4
inference-gpu[yolo-world]==0.9.12rc1
supervision==0.19.0rc3
git+https://github.com/facebookresearch/segment-anything.git
opencv-python~=4.8.0.76
numpy~=1.25.2
open3d~=0.18.0
matplotlib~=3.5.1